requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
mando==0.7.1
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
pycparser==3.0
PyGithub==2.8.1
PyJWT==2.12.1
PyNaCl==1.6.2
python-dotenv==1.2.2
radon==6.0.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
lizard
